Add User option toolbar. Move code to file. Fix query not getting the latest messages. Add new socket for when a new member enters the room.
