+++ src/components/messageDashboard/index.tsx
@@ -1,15 +1,5 @@
 import React, { useCallback, useState } from "react";
-import {
-  AppBar,
-  Avatar,
-  Box,
-  IconButton,
-  Menu,
-  MenuItem,
-  TextField,
-  Toolbar,
-  Typography,
-} from "@mui/material";
+import { Box, TextField } from "@mui/material";
 
 import { LoadingButton } from "@mui/lab";
 
@@ -17,9 +7,9 @@ import useMessages from "@customHooks/useMessages";
 import useSocket from "@customHooks/useSocket";
 import { User } from "@customTypes/users";
 import useUsers from "@customHooks/useUser";
-import useLogin from "@customHooks/useLogin";
 
 import UserMessage from "./UserMessage";
+import UserOptions from "./UserOptions";
 
 type IMessageDashboard = {
   user: User;
@@ -27,17 +17,14 @@ type IMessageDashboard = {
 
 export default function MessageDashboard(props: IMessageDashboard) {
   const { user } = props;
-  const { logout } = useLogin();
 
   const { socket } = useSocket();
   const {
     messages,
     mutate: { mutate: addMessage, isLoading: isAddingMessage },
   } = useMessages(socket);
-  const { data: users } = useUsers();
+  const { data: users } = useUsers(socket);
   const [messageText, setMessageText] = useState<string>("");
-  const [menuOpen, setMenuOpen] = useState<boolean>(false);
-  const [anchor, setAnchor] = useState<HTMLElement | undefined>();
 
   const handleSubmit = useCallback(
     async (evt: React.FormEvent<HTMLButtonElement>) => {
@@ -51,33 +38,9 @@ export default function MessageDashboard(props: IMessageDashboard) {
     [messageText, user, addMessage]
   );
 
-  const handleMenu = (evt: React.MouseEvent<HTMLButtonElement>) => {
-    setAnchor(evt.currentTarget);
-    setMenuOpen(true);
-  };
-
-  const handleClose = () => {
-    setAnchor(undefined);
-    setMenuOpen(false);
-  };
-
   return (
     <Box marginBottom="75px">
-      <AppBar position="sticky">
-        <Toolbar>
-          <Typography>Messenger App</Typography>
-          <IconButton
-            sx={{ marginLeft: "auto", color: "white" }}
-            onClick={handleMenu}
-          >
-            <Typography sx={{ marginRight: "8px" }}>{user.userName}</Typography>
-            <Avatar src={user.avatar} />
-          </IconButton>
-        </Toolbar>
-        <Menu open={menuOpen} anchorEl={anchor} onClose={handleClose}>
-          <MenuItem onClick={logout}>Logout</MenuItem>
-        </Menu>
-      </AppBar>
+      <UserOptions user={user} />
       {users && (
         <Box sx={{ overflowX: "scroll", marginTop: "8px" }}>
           {messages.map((message) => (
