+++ src/types/socket.ts
@@ -6,6 +6,7 @@ import type { NextApiResponse } from "next/types";
 export const SOCKET_EVENTS: Record<string, string> = {
   USER_MESSAGE: "user message",
   NEW_MESSAGE_UPDATE: "new message update",
+  NEW_USER: "new user",
 };
 
 export type NextSocketResponse = {
